requirements.txt
aiogrpc==1.6
anki-vector==0.6.0
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
Flask==1.0.3
googleapis-common-protos==1.6.0
grpcio==1.21.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
lib==3.0.0
MarkupSafe==1.1.1
numpy==1.20.0
Pillow==6.0.0
protobuf==3.8.0
pycparser==2.19
PyOpenGL==3.1.0
requests==2.22.0
six==1.12.0
urllib3==1.25.3
Werkzeug==0.15.4
opencv-python==4.1.2.30
pygame==1.9.6
Matplotlib ==3.5.1
